+++ rotkehlchen/api/rest.py
@@ -2794,6 +2794,14 @@ class RestAPI():
         return api_response(OK_RESULT, status_code=HTTPStatus.OK)
 
     @require_loggedin_user()
+    def get_connected_locations(self) -> Response:
+        locations = self.rotkehlchen.data.db.get_known_locations()
+        return api_response(
+            result=_wrap_in_ok_result([str(location) for location in locations]),
+            status_code=HTTPStatus.OK,
+        )
+
+    @require_loggedin_user()
     def purge_ethereum_transaction_data(self) -> Response:
         DBEthTx(self.rotkehlchen.data.db).purge_ethereum_transaction_data()
         return api_response(OK_RESULT, status_code=HTTPStatus.OK)