+++ rotkehlchen/api/server.py
@@ -67,12 +67,13 @@ from rotkehlchen.api.v1.resources import (
     IgnoredActionsResource,
     IgnoredAssetsResource,
     InfoResource,
+    KnownLocations,
     LedgerActionsResource,
     LimitsCounterResetResource,
-    LiquityTrovesResource,
-    LiquityTrovesHistoryResource,
     LiquityStakingHistoryResource,
     LiquityStakingResource,
+    LiquityTrovesHistoryResource,
+    LiquityTrovesResource,
     LoopringBalancesResource,
     MakerdaoDSRBalanceResource,
     MakerdaoDSRHistoryResource,
@@ -156,6 +157,7 @@ URLS_V1: URLS = [
     ('/exchanges/binance/pairs/<string:name>', BinanceUserMarkets),
     ('/exchanges/data/', ExchangesDataResource),
     ('/exchanges/data/<string:location>', ExchangesDataResource, 'named_exchanges_data_resource'),
+    ('/exchanges/connected/', KnownLocations),
     ('/balances/blockchains', BlockchainBalancesResource),
     (
         '/balances/blockchains/<string:blockchain>',
@@ -335,7 +337,7 @@ class APIServer():
         self.ws_server: Optional[WebSocketServer] = None
 
         self.flask_app.errorhandler(HTTPStatus.NOT_FOUND)(endpoint_not_found)  # type: ignore
-        self.flask_app.register_error_handler(Exception, self.unhandled_exception)
+        self.flask_app.register_error_handler(Exception, self.unhandled_exception)  # type: ignore
 
     @staticmethod
     def unhandled_exception(exception: Exception) -> Response: